+++ React/tic-tac/src/app/components/Header/styles.scss
@@ -56,7 +56,6 @@ body {
   box-shadow: 0 4px 0 $bg-dark-green;
   color: #FFF;
   cursor: pointer;
-  float: right;
   font-size: 14px;
   font-weight: bold;
   margin: 15px;
