+++ React/tic-tac/src/app/screens/Game/style.scss
@@ -2,10 +2,6 @@
 @import 'fonts';
 @import 'backgrounds';
 
-body {
-  margin: 0;
-}
-
 .game {
   display: flex;
   margin: 0 auto;
@@ -54,33 +50,6 @@ body {
 
   &:hover {
     background-color: $bg-green;
-  }
-}
-
-.app-logo {
-  animation: app-logo-spin infinite 20s linear;
-  height: 100px;
-}
-
-.app-header {
-  background-color: $bg-gray-light;
-  height: auto;
-  padding: 20px;
-  text-align: center;
-}
-
-.app-title {
-  color: $color-gray;
-  font-family: $font-cursive;
-  font-size: 35px;
-}
-
-@keyframes app-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
   }
 }
