+++ React/tic-tac/src/app/screens/Login/components/formField/style.scss
@@ -1,0 +1,10 @@
+.alert {
+  background-color: pink;
+  border: 2px solid #CCC;
+  display: block;
+  font-size: 12px;
+}
+
+.field {
+  display: block;
+}
